iss3 - Create a webservice
We need to be able to export and import questions to/from Moodle via the command line
on a user's own computer.
The most secure way of doing this seems to be using
a webservice. We can access using a token e.g.

http://STACK_URL/webservice/rest/server.php?wstoken=XXXX&wsfunction=qbank_gitsync_export_question&moodlewsrestformat=json&questionid=20

This is a basic proof of concept of getting questions in and out. We also need to handle metadata, etc.