recovery - Deal with repo creation via import
Update the process to handle the user starting with a repo and then importing it to Moodle rather than creating the repo from Moodle.

Also handle what happens when questions are added to the repo.

Requires update of commit hashes during main processing, not just from the CLIs.

Add 'usegit' config to switch this on and off.